+++ src/App.vue
@@ -1,11 +1,3 @@
 <template>
-  <div class="justify-center flex-1">
-    <div class="container pt-8 mx-auto">
-      <HelloWorld msg="Hello Vue 3 + Vite + Tailwind CSS" />
-    </div>
-  </div>
+  <router-view></router-view>
 </template>
-
-<script setup lang="ts">
-import HelloWorld from './components/HelloWorld.vue'
-</script>