+++ src/components/HelloWorld.vue
@@ -1,25 +1,22 @@
 <template>
-  <div class="px-8 prose">
-    <h1>{{ msg }}</h1>
+  <div>
+    <h2>{{ msg }}</h2>
     <button class="px-3 py-2 border border-gray-300 rounded-md shadow" @click="increment">
       count is: {{ count }}
     </button>
     <p>Edit <code>components/HelloWorld.vue</code> to test hot module replacement.</p>
-
-    <p class="text-sm text-gray-700">vite-ts-tailwind-starter v{{ appVersion }}</p>
   </div>
 </template>
 
 <script setup lang="ts">
-import { computed, defineProps } from 'vue'
-
+import { computed } from 'vue'
 import { useStore, Mutation } from '@/store/index'
 
 const store = useStore()
-const appVersion = store.state.version // not reactive!
 const count = computed(() => store.state.count)
 const props = defineProps<{
   msg: string
+  optionalProp?: number
 }>()
 
 function increment() {