+++ src/env.d.ts
@@ -1,0 +1,4 @@
+interface ImportMetaEnv {
+  VITE_APP_VERSION: string
+  VITE_APP_BUILD_EPOCH: number
+}