+++ src/pages/About.vue
@@ -1,0 +1,7 @@
+<template>
+  <BaseTemplate>
+    <p>I'm just a dumb about page with no content.</p>
+
+    <p><router-link to="/">Back to homepage</router-link></p>
+  </BaseTemplate>
+</template>