+++ src/shims-vue.d.ts
@@ -3,5 +3,3 @@ declare module '*.vue' {
   const _default: Component
   export default _default
 }
-
-declare const _APP_VERSION: string