+++ src/store/state.ts
@@ -5,7 +5,10 @@ export interface State {
   count: number
 }
 
-const versionString = import.meta.env.MODE === 'development' ? _APP_VERSION + '-dev' : _APP_VERSION
+const versionString =
+  import.meta.env.MODE === 'development'
+    ? import.meta.env.VITE_APP_VERSION + '-dev'
+    : import.meta.env.VITE_APP_VERSION
 
 export const state: State = {
   debug: import.meta.env.MODE === 'development',