+++ vite.config.js
@@ -3,14 +3,14 @@ import { defineConfig } from 'vite'
 import vue from '@vitejs/plugin-vue'
 import pkg from './package.json'
 
+process.env.VITE_APP_BUILD_EPOCH = new Date().getTime()
+process.env.VITE_APP_VERSION = pkg.version
+
 export default defineConfig({
   plugins: [vue()],
   resolve: {
     alias: {
       '@': path.resolve(__dirname, './src'),
     },
-  },
-  define: {
-    _APP_VERSION: JSON.stringify(pkg.version),
   },
 })